Split check.py into module.
